refactor(web): authenticate機能のui.tsバレルを削除して構成を整理
他のfeature（article/diary/inbox）と同様にindex.tsから
components配下を直接re-exportする形に揃え、冗長な中間バレルを除去する。
