+++ application/packages/web/src/client/components/ui/layout/sidebar/index.tsx
@@ -10,7 +10,7 @@ import {
 } from '@/client/components/shadcn/sidebar'
 import { AnchorLink } from '@/client/components/ui/navigation/link'
 import NavMenu from '@/client/components/ui/navigation/nav-menu'
-import { LogoutButton } from '@/client/features/authenticate/ui'
+import { LogoutButton } from '@/client/features/authenticate'
 import type { InternalPath } from '@/client/routes'
 
 export interface MenuItem {
